requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
griffe==1.12.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
mcp==1.13.1
ollama==0.5.3
openai==1.101.0
openai-agents==0.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
setuptools==80.9.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.3
tqdm==4.67.1
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wheel==0.45.1
